test y contenido completo

+++ database/migrations/2025_06_14_022360_create_questions_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('questions', function (Blueprint $table) {
             $table->id();
             $table->string('statement');
-            $table->decimal('spot', 2, 1)->default(1);
+            $table->integer('spot')->default(1);
             $table->unsignedBigInteger('type_questions_id');
             $table->unsignedBigInteger('chapter_id');
             $table->timestamps();
